+++ src/test/java/com/nylas/NylasAccountTest.java
@@ -5,10 +5,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.util.Collections;
+import java.util.Map;
 
 import static com.nylas.AccessTokenTest.TEST_ACCESS_TOKEN;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
@@ -61,11 +62,42 @@ public class NylasAccountTest {
     @Test
     public void testRevokeAccessToken() throws RequestFailedException, IOException {
         final NylasAccount nylasAccount = new NylasAccount(nylasClient, TEST_ACCESS_TOKEN);
+        final Map<String, Boolean> revokeResponse = Collections.singletonMap("success", true);
 
         when(nylasClient.newUrlBuilder()).thenReturn(new HttpUrl.Builder());
+        when(nylasClient.executePost(anyString(), any(), any(), any())).thenReturn(revokeResponse);
 
-        nylasAccount.revokeAccessToken();
+        boolean revokeAccessToken = nylasAccount.revokeAccessToken();
 
         verify(nylasClient).executePost(anyString(), any(), any(), any());
+        assertTrue(revokeAccessToken);
+    }
+
+    @Test
+    public void testRevokeAccessTokenSuccessFalse() throws RequestFailedException, IOException {
+        final NylasAccount nylasAccount = new NylasAccount(nylasClient, TEST_ACCESS_TOKEN);
+        final Map<String, Boolean> revokeResponse = Collections.singletonMap("success", false);
+
+        when(nylasClient.newUrlBuilder()).thenReturn(new HttpUrl.Builder());
+        when(nylasClient.executePost(anyString(), any(), any(), any())).thenReturn(revokeResponse);
+
+        boolean revokeAccessToken = nylasAccount.revokeAccessToken();
+
+        verify(nylasClient).executePost(anyString(), any(), any(), any());
+        assertFalse(revokeAccessToken);
+    }
+
+    @Test
+    public void testRevokeAccessTokenInvalidObjectReturnsFalse() throws RequestFailedException, IOException {
+        final NylasAccount nylasAccount = new NylasAccount(nylasClient, TEST_ACCESS_TOKEN);
+        final Map<String, Object> revokeResponse = Collections.singletonMap("invalidKey", "invalidValue");
+
+        when(nylasClient.newUrlBuilder()).thenReturn(new HttpUrl.Builder());
+        when(nylasClient.executePost(anyString(), any(), any(), any())).thenReturn(revokeResponse);
+
+        boolean revokeAccessToken = nylasAccount.revokeAccessToken();
+
+        verify(nylasClient).executePost(anyString(), any(), any(), any());
+        assertFalse(revokeAccessToken);
     }
 }